requirements.txt
certifi
cycler
future
joblib
kiwisolver
matplotlib
numpy
opencv-python
Pillow
protobuf
pyparsing
python-dateutil
scikit-learn
scipy
six
tensorboard
tensorboardX
threadpoolctl
torch
torchvision
tqdm

PyYAML